misc/telegram_bot/requirements.txt
anyio==4.1.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
exceptiongroup==1.2.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0
gspread==5.12.3
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
oauthlib==3.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
python-telegram-bot==20.7
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
sniffio==1.3.0
urllib3==2.1.0
